feat/#25: HomeScreen 내 컴포넌트 파일 분리
